+++ src/users_groups_lib/errors/__init__.py
@@ -1,3 +1,3 @@
 """Exposed errors classes and methods."""
 from users_groups_lib.errors.user import UserExistError, UserPermissionError, UserNotExistError, UserInUseError
-from users_groups_lib.errors.group import GroupNotExistError, GroupExistError, GroupPermissionError
+from users_groups_lib.errors.group import GroupNotExistError, GroupExistError, GroupPermissionError, GroupInUseError