+++ src/users_groups_lib/managers/inserters/__init__.py
@@ -1,2 +1,3 @@
 """Exposed inserters classes and methods."""
 from users_groups_lib.managers.inserters.user_inserter import UserInserter
+from users_groups_lib.managers.inserters.group_inserter import GroupInserter